requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types @ file:///home/conda/feedstock_root/build_artifacts/annotated-types_1685597160053/work
anyio @ file:///home/conda/feedstock_root/build_artifacts/anyio_1688651106312/work/dist
argon2-cffi @ file:///opt/conda/conda-bld/argon2-cffi_1645000214183/work
argon2-cffi-bindings @ file:///Users/ec2-user/ci_py311/argon2-cffi-bindings_1678317076583/work
async-timeout @ file:///Users/ec2-user/ci_py311/async-timeout_1678320112070/work
attrs @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_8d7cifg9jt/croot/attrs_1695717834289/work
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
Bottleneck @ file:///Users/ec2-user/ci_py311/bottleneck_1678322312632/work
brotlipy==0.7.0
cchardet @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_b8vojp2a2e/croot/cchardet_1683047090189/work
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_2emnmf1qow/croot/certifi_1690232232440/work/certifi
cffi @ file:///Users/ec2-user/ci_py311/cffi_1678315312775/work
chardet @ file:///Users/ec2-user/ci_py311/chardet_1678326102860/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
chroma-hnswlib==0.7.3
chromadb==0.4.14
click @ file:///Users/ec2-user/ci_py311/click_1678318124837/work
colorama @ file:///Users/ec2-user/ci_py311/colorama_1678320227414/work
coloredlogs==15.0.1
contourpy @ file:///Users/ec2-user/ci_py311/contourpy_1678322361099/work
cryptography @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_19sc_35nzq/croot/cryptography_1694211574676/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
dataclasses-json==0.5.14
docarray==0.32.1
ecdsa @ file:///tmp/build/80754af9/ecdsa_1622554921100/work
et-xmlfile==1.1.0
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1692026125334/work
fastapi @ file:///home/conda/feedstock_root/build_artifacts/fastapi_1693793259089/work
filelock==3.12.4
Flask @ file:///Users/ec2-user/ci_py311/flask_1678380788105/work
flatbuffers==23.5.26
fonttools==4.25.0
frozenlist==1.4.0
fsspec==2023.9.2
future @ file:///Users/ec2-user/ci_py311_2/future_1679335881731/work
gpt-index==0.8.42
greenlet @ file:///Users/ec2-user/ci_py311/greenlet_1678323340336/work
grpcio==1.59.0
h11 @ file:///tmp/build/80754af9/h11_1620423447028/work
hnswlib==0.7.0
httptools==0.6.0
huggingface-hub==0.17.3
humanfriendly==10.0
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1663625384323/work
importlib-resources==6.1.0
itsdangerous @ file:///tmp/build/80754af9/itsdangerous_1621432558163/work
Jinja2 @ file:///Users/ec2-user/ci_py311/jinja2_1678317138271/work
joblib @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_dfriz0qjwt/croot/joblib_1685113297703/work
jsonpatch==1.33
jsonpointer==2.4
kiwisolver @ file:///Users/ec2-user/ci_py311/kiwisolver_1678322457192/work
langchain==0.0.312
langsmith==0.0.43
llama-index==0.8.43.post1
markdown-it-py==3.0.0
MarkupSafe @ file:///Users/ec2-user/ci_py311/markupsafe_1678316973889/work
marshmallow==3.20.1
matplotlib @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_42_ot0zpzy/croot/matplotlib-suite_1693812472014/work
mdurl==0.1.2
mkl-fft @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_af8wl3mjfu/croot/mkl_fft_1695058185839/work
mkl-random @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_0fq2s894lk/croot/mkl_random_1695059822529/work
mkl-service==2.4.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
munkres==1.1.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
nltk==3.8.1
numexpr @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_c2pokmr868/croot/numexpr_1696515300257/work
numpy @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_b1_7dulbi4/croot/numpy_and_numpy_base_1695830449184/work/dist/numpy-1.26.0-cp311-cp311-macosx_10_9_x86_64.whl#sha256=df7810f057779b4c06be5bdbd7924c5dc5b91a218ad44586e9893864b3ebaae4
onnxruntime==1.16.1
openai @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_08qwz48h3g/croot/openai_1682522446888/work
openpyxl==3.0.10
orjson==3.9.8
overrides==7.4.0
packaging==23.2
pandas @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_dflcdcqtvf/croot/pandas_1692297394605/work
pandas-stubs @ file:///Users/ec2-user/ci_py311/pandas-stubs_1678419254720/work
passlib==1.7.4
Pillow @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_18z8l26k8d/croot/pillow_1696580044342/work
plotly @ file:///Users/ec2-user/ci_py311/plotly_1678382250051/work
posthog==3.0.2
protobuf==4.24.4
psycopg2-binary==2.9.7
pulsar-client==3.3.0
pyasn1 @ file:///Users/ktietz/demo/mc3/conda-bld/pyasn1_1629708007385/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodome @ file:///Users/ec2-user/ci_py311/pycryptodome_1678382810138/work
pydantic @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_92rc5e4spx/croot/pydantic_1695798855907/work
Pygments==2.16.1
PyMySQL @ file:///Users/ec2-user/ci_py311/pymysql_1678420708564/work
pyOpenSSL @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_7fwdg30481/croot/pyopenssl_1690223429070/work
pyparsing @ file:///Users/ec2-user/ci_py311/pyparsing_1678322828710/work
PyPika==0.48.9
PySocks @ file:///Users/ec2-user/ci_py311/pysocks_1678315868424/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-dotenv==1.0.0
python-jose @ file:///tmp/build/80754af9/python-jose_1623673016654/work
python-multipart @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_2ds_a7enty/croot/python-multipart_1692946343582/work
pytz @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_e0y2h1cdsw/croot/pytz_1695131602326/work
PyYAML==6.0.1
regex==2023.10.3
requests @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_fa2zrx4csg/croot/requests_1690400223595/work
rich==13.6.0
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
scikit-learn @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_815xfdnpoa/croot/scikit-learn_1680198756636/work
scipy @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_d8wk4_24cp/croot/scipy_1696543301467/work/dist/scipy-1.11.3-cp311-cp311-macosx_10_9_x86_64.whl#sha256=b00f65c1d01b59ce13a96cf994acd9d6076afacd866199dd3bfa980964b8936a
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio @ file:///home/conda/feedstock_root/build_artifacts/sniffio_1662051266223/work
soupsieve==2.5
SQLAlchemy==2.0.21
starlette @ file:///home/conda/feedstock_root/build_artifacts/starlette-recipe_1684245096404/work
sympy==1.12
tenacity==8.2.3
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
tiktoken==0.5.1
tokenizers==0.14.1
tqdm @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_03acd6t6ca/croot/tqdm_1679561866522/work
typer==0.9.0
types-pytz @ file:///Users/ec2-user/ci_py311/types-pytz_1678394832420/work
types-requests==2.31.0.8
typing-inspect==0.9.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1695040754690/work
tzdata @ file:///croot/python-tzdata_1690578112552/work
urllib3==1.26.17
uvicorn @ file:///Users/ec2-user/ci_py311_2/uvicorn-split_1679338468408/work
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
Werkzeug @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_fegt3z60s3/croot/werkzeug_1679489762631/work
yarl @ file:///Users/ec2-user/ci_py311/yarl_1678323149735/work
